feat: add empty fav page component
