+++ job-search/src/components/EmptyFavorites/hooks.ts
@@ -1,0 +1,13 @@
+import { useNavigate } from 'react-router-dom'
+
+import { PATHS } from 'router/paths'
+
+export const useEmptyFavorites = () => {
+  const navigate = useNavigate()
+
+  const handleBtnClick = () => {
+    navigate(PATHS.VACANCIES)
+  }
+
+  return { handleBtnClick }
+}
