+++ job-search/src/components/EmptyFavorites/index.ts
@@ -1,0 +1,1 @@
+export { default } from './EmptyFavorites'
